requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.19.1
fsspec==2025.7.0
grpcio==1.74.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.7
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.22.3
packaging==25.0
pillow==11.3.0
portalocker==3.2.0
protobuf==6.32.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
qdrant-client==1.15.1
regex==2025.7.34
requests==2.32.4
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.11.4
sentence-transformers==2.7.0
sniffio==1.3.1
starlette==0.47.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==1.13.1
tqdm==4.67.1
transformers==4.41.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0